mlservice/requirements.txt
fastapi
uvicorn[standard]
torch
open-clip-torch
mobileclip
numpy
Pillow
